This is essentially the "last" update to this little engine that could(n't even).
I want this to stay in it's current form for awhile, while I go forward with new branches, and try to test different genres/styles of games, to see how it handles them/how it can be improved in a general sense.

THIS COMMIT ADDED:
-Lighting/shadows are now 95% done. There are a few things that I would've liked to try/change, but that seems a little too specific for this "general" base engine. (And most of those things should be in "general bug list.txt".
-And that's kinda it, I think. There were additions/design changes here and there, but all were generally small and would take ages to explain each. Just lots of little additions.

Now, GO FORTH! And start iterating on all the knowledge/design elements you've learned from this!