This is a DEFINITE work-in-progress/test build of the engine.
Working on projectiles/collision detection here, but only using tile-based collision. I want to delve in deeper than that, so my code is going to be not-so-organized for awhile, while I find and figure out what collision method I'd like to use. This commit is just being used to save my last "version" of the engine, that I myself have been playing around with.

That's all. This is more or less just a fallback commit, in case things go horribly wrong here on out.